src/requirements.txt
pyjwt
psycopg2-binary==2.9.3
